updated the code so api call functions are in index.js file and custom hooks are in a separate file
